Move the code which launches JVM to a separate class
